chore: address issue #152
